Refactor Database panel UI
