+++ packages/playground/website/src/components/site-manager/site-database-panel/phpmyadmin-button.tsx
@@ -1,0 +1,105 @@
+import { useState } from 'react';
+import { Button } from '@wordpress/components';
+import {
+	type PlaygroundClient,
+	type StepDefinition,
+	type UniversalPHP,
+	compileBlueprintV1,
+	runBlueprintV1Steps,
+} from '@wp-playground/client';
+// @ts-ignore
+import { corsProxyUrl } from 'virtual:cors-proxy-url';
+
+const phpMyAdminUrl =
+	'https://files.phpmyadmin.net/phpMyAdmin/5.2.3/phpMyAdmin-5.2.3-english.zip';
+
+async function installPhpMyAdmin(playground: PlaygroundClient) {
+	const documentRoot = await playground.documentRoot;
+	const phpMyAdminPath = `${documentRoot}/phpmyadmin`;
+
+	const steps: StepDefinition[] = [
+		{
+			step: 'unzip',
+			zipFile: {
+				resource: 'url',
+				url: phpMyAdminUrl,
+			},
+			extractToPath: documentRoot,
+		},
+		{
+			step: 'mv',
+			fromPath: `${documentRoot}/phpMyAdmin-5.2.3-english`,
+			toPath: phpMyAdminPath,
+		},
+		{
+			step: 'writeFile',
+			path: `${phpMyAdminPath}/libraries/classes/Dbal/DbiMysqli.php`,
+			data: (await import('./phpmyadmin-extensions/DbiMysqli.php?raw'))
+				.default as string,
+		},
+		{
+			step: 'writeFile',
+			path: `${phpMyAdminPath}/config.inc.php`,
+			data: (await import('./phpmyadmin-extensions/config.inc.php?raw'))
+				.default as string,
+		},
+	];
+
+	const blueprint = await compileBlueprintV1(
+		{ steps },
+		{ corsProxy: corsProxyUrl }
+	);
+
+	await runBlueprintV1Steps(blueprint, playground as UniversalPHP);
+}
+
+export function PhpMyAdminButton({
+	playground,
+}: {
+	playground: PlaygroundClient | undefined;
+}) {
+	const [state, setState] = useState<'idle' | 'loading' | 'ready'>('idle');
+
+	const handleOpenPhpMyAdmin = async () => {
+		if (!playground) {
+			return;
+		}
+
+		if (state === 'loading') {
+			return;
+		}
+
+		if (state === 'idle') {
+			setState('loading');
+			try {
+				await installPhpMyAdmin(playground);
+				setState('ready');
+			} catch (error) {
+				console.error('Failed to install phpMyAdmin:', error);
+				setState('idle');
+				return;
+			}
+		}
+
+		const playgroundUrl = await playground.absoluteUrl;
+		if (playgroundUrl) {
+			window.open(
+				`${playgroundUrl}/phpmyadmin/index.php?route=/database/structure&db=wordpress`,
+				'_blank',
+				'noopener,noreferrer'
+			);
+		}
+	};
+
+	const isLoading = state === 'loading';
+	return (
+		<Button
+			variant="primary"
+			disabled={!playground || isLoading}
+			isBusy={isLoading}
+			onClick={handleOpenPhpMyAdmin}
+		>
+			{isLoading ? 'Setting up phpMyAdmin…' : 'phpMyAdmin'}
+		</Button>
+	);
+}
